+++ apps/meteor/app/livechat/server/hooks/afterUserActions.ts
@@ -1,23 +1,51 @@
 import type { IUser } from '@rocket.chat/core-typings';
-import { Users } from '@rocket.chat/models';
-import type { UsersUpdateParamsPOST } from '@rocket.chat/rest-typings';
 
 import { callbacks } from '../../../../lib/callbacks';
+import { Livechat } from '../lib/Livechat';
+import { callbackLogger } from '../lib/logger';
 
-type UserData = UsersUpdateParamsPOST['data'] & { _id: string };
+type IAfterSaveUserProps = {
+	user: IUser;
+	oldUser: IUser | null;
+};
 
-const handleAgentUpdated = async (userData: UserData) => {
-	if (!userData?.roles?.includes('livechat-agent')) {
-		await Users.unsetExtension(userData._id);
+const wasAgent = (user: Pick<IUser, 'roles'> | null) => user?.roles?.includes('livechat-agent');
+const isAgent = (user: Pick<IUser, 'roles'> | null) => user?.roles?.includes('livechat-agent');
+
+const handleAgentUpdated = async (userData: IAfterSaveUserProps) => {
+	const {
+		user: { _id: userId, username },
+		user: newUser,
+		oldUser,
+	} = userData;
+
+	if (wasAgent(oldUser) && !isAgent(newUser)) {
+		callbackLogger.debug('Removing agent', userId);
+		await Livechat.removeAgent(username);
+	}
+
+	if (!wasAgent(oldUser) && isAgent(newUser)) {
+		callbackLogger.debug('Adding agent', userId);
+		await Livechat.addAgent(username);
 	}
 };
 
-const handleDeactivateUser = async (userData: IUser) => {
-	if (userData?.roles?.includes('livechat-agent')) {
-		await Users.unsetExtension(userData._id);
+const handleDeactivateUser = async (user: IUser) => {
+	if (wasAgent(user)) {
+		callbackLogger.debug('Removing agent', user._id);
+		await Livechat.removeAgent(user.username);
 	}
 };
 
-callbacks.add('afterSaveUser', handleAgentUpdated, callbacks.priority.LOW, 'livechat-after-save-user-remove-extension');
+const handleActivateUser = async (user: IUser) => {
+	if (isAgent(user)) {
+		callbackLogger.debug('Adding agent', user._id);
+		await Livechat.addAgent(user.username);
+	}
+};
 
-callbacks.add('afterDeactivateUser', handleDeactivateUser, callbacks.priority.LOW, 'livechat-after-deactivate-user-remove-extension');
+callbacks.add('afterSaveUser', handleAgentUpdated, callbacks.priority.LOW, 'livechat-after-save-user-update-agent');
+
+callbacks.add('afterDeactivateUser', handleDeactivateUser, callbacks.priority.LOW, 'livechat-after-deactivate-user-remove-agent');
+
+callbacks.add('afterActivateUser', handleActivateUser, callbacks.priority.LOW, 'livechat-after-activate-user-add-agent');