+++ apps/meteor/client/views/account/tokens/AccountTokensTable/AccountTokensTable.tsx
@@ -47,11 +47,11 @@ const AccountTokensTable = (): ReactElement => {
 	const headers = useMemo(
 		() =>
 			[
-				<GenericTableHeaderCell key={'name'}>{t('API_Personal_Access_Token_Name')}</GenericTableHeaderCell>,
-				isMedium && <GenericTableHeaderCell key={'createdAt'}>{t('Created_at')}</GenericTableHeaderCell>,
-				<GenericTableHeaderCell key={'lastTokenPart'}>{t('Last_token_part')}</GenericTableHeaderCell>,
-				<GenericTableHeaderCell key={'2fa'}>{t('Two Factor Authentication')}</GenericTableHeaderCell>,
-				<GenericTableHeaderCell key={'actions'} />,
+				<GenericTableHeaderCell key='name'>{t('API_Personal_Access_Token_Name')}</GenericTableHeaderCell>,
+				isMedium && <GenericTableHeaderCell key='createdAt'>{t('Created_at')}</GenericTableHeaderCell>,
+				<GenericTableHeaderCell key='lastTokenPart'>{t('Last_token_part')}</GenericTableHeaderCell>,
+				<GenericTableHeaderCell key='2fa'>{t('Two Factor Authentication')}</GenericTableHeaderCell>,
+				<GenericTableHeaderCell key='actions' />,
 			].filter(Boolean),
 		[isMedium, t],
 	);