+++ apps/meteor/ee/client/deviceManagement/components/DeviceIcon.tsx
@@ -1,8 +1,9 @@
 import { Box, Icon } from '@rocket.chat/fuselage';
-import type { ComponentProps, ReactElement } from 'react';
+import type { Keys as IconName } from '@rocket.chat/icons';
+import type { ReactElement } from 'react';
 import React from 'react';
 
-const iconMap: Record<string, ComponentProps<typeof Icon>['name']> = {
+const iconMap: Record<string, IconName> = {
 	'browser': 'desktop',
 	'mobile': 'mobile',
 	'desktop-app': 'desktop',