+++ apps/meteor/server/models/raw/Users.js
@@ -66,6 +66,18 @@ export class UsersRaw extends BaseRaw {
 			{ key: { language: 1 }, sparse: true },
 			{ key: { 'active': 1, 'services.email2fa.enabled': 1 }, sparse: true }, // used by statistics
 			{ key: { 'active': 1, 'services.totp.enabled': 1 }, sparse: true }, // used by statistics
+			// Used for case insensitive queries
+			// @deprecated
+			// Should be converted to unique index later within a migration to prevent errors of duplicated
+			// records. Those errors does not helps to identify the duplicated value so we need to find a
+			// way to help the migration in case it happens.
+			{
+				key: { 'emails.address': 1 },
+				unique: false,
+				sparse: true,
+				name: 'emails.address_insensitive',
+				collation: { locale: 'en', strength: 2, caseLevel: false },
+			},
 		];
 	}
 
@@ -842,9 +854,6 @@ export class UsersRaw extends BaseRaw {
 		};
 
 		const update = {
-			$set: {
-				statusLivechat: 'available',
-			},
 			$addToSet: {
 				openBusinessHours: { $each: businessHourIds },
 			},
@@ -860,9 +869,6 @@ export class UsersRaw extends BaseRaw {
 		};
 
 		const update = {
-			$set: {
-				statusLivechat: 'available',
-			},
 			$addToSet: {
 				openBusinessHours: { $each: businessHourIds },
 			},
@@ -878,11 +884,27 @@ export class UsersRaw extends BaseRaw {
 		};
 
 		const update = {
-			$set: {
-				statusLivechat: 'available',
-			},
 			$addToSet: {
 				openBusinessHours: businessHourId,
+			},
+		};
+
+		return this.updateMany(query, update);
+	}
+
+	makeAgentsWithinBusinessHourAvailable(agentIds) {
+		const query = {
+			...(agentIds && { _id: { $in: agentIds } }),
+			roles: 'livechat-agent',
+			// Exclude away users
+			status: 'online',
+			// Exclude users that are already available, maybe due to other business hour
+			statusLivechat: 'not-available',
+		};
+
+		const update = {
+			$set: {
+				statusLivechat: 'available',
 			},
 		};
 
@@ -910,9 +932,6 @@ export class UsersRaw extends BaseRaw {
 		};
 
 		const update = {
-			$set: {
-				statusLivechat: 'available',
-			},
 			$addToSet: {
 				openBusinessHours: businessHourId,
 			},
@@ -953,6 +972,8 @@ export class UsersRaw extends BaseRaw {
 		const query = {
 			$or: [{ openBusinessHours: { $exists: false } }, { openBusinessHours: { $size: 0 } }],
 			roles: 'livechat-agent',
+			// Avoid unnecessary updates
+			statusLivechat: 'available',
 			...(Array.isArray(userIds) && userIds.length > 0 && { _id: { $in: userIds } }),
 		};
 
@@ -968,6 +989,7 @@ export class UsersRaw extends BaseRaw {
 	setLivechatStatusActiveBasedOnBusinessHours(userId) {
 		const query = {
 			_id: userId,
+			statusDefault: { $ne: 'offline' },
 			openBusinessHours: {
 				$exists: true,
 				$not: { $size: 0 },
@@ -984,15 +1006,14 @@ export class UsersRaw extends BaseRaw {
 	}
 
 	async isAgentWithinBusinessHours(agentId) {
-		return (
-			(await this.find({
-				_id: agentId,
-				openBusinessHours: {
-					$exists: true,
-					$not: { $size: 0 },
-				},
-			}).count()) > 0
-		);
+		const query = {
+			_id: agentId,
+			openBusinessHours: {
+				$exists: true,
+				$not: { $size: 0 },
+			},
+		};
+		return (await this.col.countDocuments(query)) > 0;
 	}
 
 	removeBusinessHoursFromAllUsers() {
@@ -1254,6 +1275,16 @@ export class UsersRaw extends BaseRaw {
 
 	findOneByResetToken(token, options) {
 		return this.findOne({ 'services.password.reset.token': token }, options);
+	}
+
+	findOneByIdWithEmailAddress(userId, options) {
+		return this.findOne(
+			{
+				_id: userId,
+				emails: { $exists: true, $ne: [] },
+			},
+			options,
+		);
 	}
 
 	setFederationAvatarUrlById(userId, federationAvatarUrl) {
@@ -1891,17 +1922,23 @@ export class UsersRaw extends BaseRaw {
 
 	findOneByEmailAddressAndServiceNameIgnoringCase(emailAddress, userId, serviceName, options) {
 		const query = {
-			'emails.address': new RegExp(`^${escapeRegExp(String(emailAddress).trim())}$`, 'i'),
+			'emails.address': String(emailAddress).trim(),
 			[`services.${serviceName}.id`]: userId,
 		};
 
-		return this.findOne(query, options);
+		return this.findOne(query, {
+			collation: { locale: 'en', strength: 2 }, // Case insensitive
+			...options,
+		});
 	}
 
 	findOneByEmailAddress(emailAddress, options) {
-		const query = { 'emails.address': String(emailAddress).trim().toLowerCase() };
+		const query = { 'emails.address': String(emailAddress).trim() };
 
-		return this.findOne(query, options);
+		return this.findOne(query, {
+			collation: { locale: 'en', strength: 2 }, // Case insensitive
+			...options,
+		});
 	}
 
 	findOneWithoutLDAPByEmailAddress(emailAddress, options) {
@@ -2821,19 +2858,17 @@ export class UsersRaw extends BaseRaw {
 	// here
 	getActiveLocalUserCount() {
 		return Promise.all([
+			// Count all active users (fast based on index)
 			this.col.countDocuments({
 				active: true,
-				type: {
-					$nin: ['app'],
-				},
-				roles: { $ne: ['guest'] },
 			}),
-			this.col.countDocuments({ federated: true, active: true }),
+			// Count all active that are guests, apps or federated
+			// Fast based on indexes, usually based on guest index as is usually small
 			this.col.countDocuments({
-				isRemote: true,
 				active: true,
-				roles: { $ne: ['guest'] },
+				$or: [{ roles: ['guest'] }, { type: 'app' }, { federated: true }, { isRemote: true }],
 			}),
+			// Get all active and remove the guests, apps, federated, etc
 		]).then((results) => results.reduce((a, b) => a - b));
 	}
 
@@ -2885,5 +2920,21 @@ export class UsersRaw extends BaseRaw {
 
 	countRoomMembers(roomId) {
 		return this.col.countDocuments({ __rooms: roomId, active: true });
+	}
+
+	removeAgent(_id) {
+		const update = {
+			$set: {
+				operator: false,
+			},
+			$unset: {
+				livechat: 1,
+				statusLivechat: 1,
+				extension: 1,
+				openBusinessHours: 1,
+			},
+		};
+
+		return this.updateOne({ _id }, update);
 	}
 }