+++ ee/apps/authorization-service/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @rocket.chat/authorization-service
 
+## 0.1.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.2.10
+- @rocket.chat/rest-typings@6.2.10
+- @rocket.chat/core-services@0.0.5
+- @rocket.chat/model-typings@0.0.5
+- @rocket.chat/models@0.0.5
+
+## 0.1.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.2.9
+- @rocket.chat/rest-typings@6.2.9
+- @rocket.chat/core-services@0.0.4
+- @rocket.chat/model-typings@0.0.4
+- @rocket.chat/models@0.0.4
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @rocket.chat/core-typings@6.2.7
+  - @rocket.chat/rest-typings@6.2.7
+  - @rocket.chat/core-services@0.0.3
+  - @rocket.chat/model-typings@0.0.3
+  - @rocket.chat/models@0.0.3
+
 ## 0.1.1
 
 ### Patch Changes