+++ packages/instance-status/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @rocket.chat/instance-status
 
+## 0.0.5
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- @rocket.chat/models@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @rocket.chat/models@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes